Run trace test on Native Driver if they don't reach 60 on ANGLE
On some low-end devices, trace tests can't reach 60 fps on Native
GLES driver. It doesn't make sense to require these trace tests
to reach 60 fps on ANGLE. This change enables the trace to also
run with native driver if it fails to reach 60 fps. If the trace
also fails to reach 60 fps, we mark this trace test passes, as
ANGLE does not make the trace perform worse.

Flag: TEST_ONLY
Test: atest AngleAllowlistTraceTest
Bug: b/331266802
Bug: b/394587449
Change-Id: Ie4887486ff5caf9b975e0601e5e73fcf7381bc91