Move the test requirement check in setUp()
The setUp() step is checking if the test asset has been
downloaded as stated in the instruction in
https://docs.partner.android.com/gms/testing/eap/vts-eap#test_setup_instruction_for_angleallowlisttracetest.

However if the device does not meet test requirement,
we should not enforce the test asset download step is completed.

Move the test requirement check in the setUp() so that the test asset
check will be skipped if the device does not meet the test requirement.

Flag: TEST_ONLY
Test: atest AngleAllowlistTraceTest
Bug: b/409750970
Change-Id: Ib1540332aa9c732bfb34c0840da8bdd722e3a9c8