Skip AngleAllowlistTraceTest on devices where ANGLE is native driver
On these devices, traces already run with ANGLE by default.
There is no point comparing trace performance with ANGLE and
with Native driver as they are the same driver.

Also remove the logging of uninstallAppIgnoreErrors exceptions.
We are ignoring the exception anyway, and printing the exception
message in test logs will create confusion: it looks like test
failure callstack but in reality that callstack does not cause
test to fail.

Flag: TEST_ONLY
Test: atest AngleAllowlistTraceTest
Bug: b/461415295
Change-Id: I909c8e19b56b3246be24fd50c9cdc5a134964fc0